fix(api): move unauthenticated helpers out of "use server" module
fetchUserCostDataInternal and fetchProfileDataInternal were exported
from a "use server" file, making them callable as Server Actions from
client components without auth checks. Moved to src/lib/profile-data.ts
which has no "use server" directive.
